+++ guillotina/interfaces/response.py
@@ -1,3 +1,5 @@
+from multidict import CIMultiDict
+from typing import Union
 from zope.interface import Attribute
 from zope.interface import Interface
 
@@ -7,12 +9,18 @@ class IResponse(Interface):
     content = Attribute("content")
     headers = Attribute("headers")
 
-    def __init__(*, content: dict = None, headers: dict = None, status: int = None):
+    def __init__(
+        *,
+        body: bytes = None,
+        content: dict = None,
+        headers: Union[dict, CIMultiDict] = None,
+        status: int = None,
+        content_type: str = None,
+        content_length: int = None,
+    ):
         """
         """
 
-
-class IASGIResponse(Interface):
-    """
-    Mark ASGI server responses with interface
-    """
+    def set_body(body):
+        """
+        """